chore: pre-flight fixes (tc-9oyhw)
swift-format style tweaks to the files this bead touches (Text() wrapping
over 120 cols; brace-on-new-line reverted back to same-line where it
conflicted with SwiftLint's opening_brace rule -- SwiftLint is the arbiter
here), and splits the new 400/409 watch-zone cases out of
URLSessionAPIClientTests into a dedicated
URLSessionAPIClientWatchZoneErrorMappingTests -- the combined file tripped
SwiftLint's file_length rule under --strict, mirroring the existing
URLSessionAPIClientTokenRefreshTests/URLSessionAPIClientPagedTests split.

swift-format --recursive also reformatted ~40 unrelated files across the
repo; those are reverted here to keep this PR's diff scoped to tc-9oyhw.

swiftlint lint --strict: 0 violations. swift test: 1979/1979 passed.
