+++ mobile/ios/town-crier-tests/Sources/Features/URLSessionAPIClientTests.swift
@@ -260,97 +260,6 @@ struct URLSessionAPIClientTests {
     }
   }
 
-  // MARK: - 400/409 watch-zone error mapping (tc-9oyhw, GH#1085)
-
-  @Test("400 with boundary_too_large body throws DomainError.invalidWatchZoneBoundaryTooLarge")
-  func boundaryTooLargeResponse() async throws {
-    let transport = StubHTTPTransport()
-    let authService = SpyAuthenticationService()
-    authService.currentSessionResult = .valid
-
-    let body =
-      #"{"error":"boundary_too_large","message":"The boundary is too large for your subscription tier."}"#
-    transport.responses = [
-      (Data(body.utf8), httpResponse(url: baseURL, statusCode: 400))
-    ]
-
-    let sut = URLSessionAPIClient(
-      baseURL: baseURL,
-      authService: authService,
-      transport: transport
-    )
-
-    await #expect(throws: DomainError.invalidWatchZoneBoundaryTooLarge) {
-      let _: TestResponse = try await sut.request(.post("/watch-zones", body: TestBody(title: "x")))
-    }
-  }
-
-  @Test("409 with zone_name_taken body throws DomainError.watchZoneNameTaken")
-  func zoneNameTakenResponse() async throws {
-    let transport = StubHTTPTransport()
-    let authService = SpyAuthenticationService()
-    authService.currentSessionResult = .valid
-
-    let body = #"{"error":"zone_name_taken","message":"You already have a watch zone with this name."}"#
-    transport.responses = [
-      (Data(body.utf8), httpResponse(url: baseURL, statusCode: 409))
-    ]
-
-    let sut = URLSessionAPIClient(
-      baseURL: baseURL,
-      authService: authService,
-      transport: transport
-    )
-
-    await #expect(throws: DomainError.watchZoneNameTaken) {
-      let _: TestResponse = try await sut.request(.post("/watch-zones", body: TestBody(title: "x")))
-    }
-  }
-
-  @Test("400 with unrecognized error code falls back to APIError.serverError")
-  func unrecognizedBadRequestFallsBackToServerError() async throws {
-    let transport = StubHTTPTransport()
-    let authService = SpyAuthenticationService()
-    authService.currentSessionResult = .valid
-
-    let body = #"{"error":"validation_failed","message":"Something else is wrong."}"#
-    transport.responses = [
-      (Data(body.utf8), httpResponse(url: baseURL, statusCode: 400))
-    ]
-
-    let sut = URLSessionAPIClient(
-      baseURL: baseURL,
-      authService: authService,
-      transport: transport
-    )
-
-    await #expect(throws: APIError.self) {
-      let _: TestResponse = try await sut.request(.post("/watch-zones", body: TestBody(title: "x")))
-    }
-  }
-
-  @Test("409 with malformed body falls back to APIError.serverError")
-  func malformedConflictFallsBackToServerError() async throws {
-    let transport = StubHTTPTransport()
-    let authService = SpyAuthenticationService()
-    authService.currentSessionResult = .valid
-
-    let body = "not json"
-    transport.responses = [
-      (Data(body.utf8), httpResponse(url: baseURL, statusCode: 409))
-    ]
-
-    let sut = URLSessionAPIClient(
-      baseURL: baseURL,
-      authService: authService,
-      transport: transport
-    )
-
-    await #expect(throws: APIError.self) {
-      let _: TestResponse = try await sut.request(.post("/watch-zones", body: TestBody(title: "x")))
-    }
-  }
-
   // MARK: - Network errors
 
   @Test("Network error maps to DomainError.networkUnavailable")
